fix edit announcement, edit sub category, and expired token

+++ src/service/SubCategoryService.ts
@@ -1,5 +1,6 @@
 import { SubCategory, SubCategorySaveDTO } from "@/types/subCategory";
 import AxiosService from "./AxiosService";
+import TokenService from "./TokenService";
 
 class SubCategoryService {
     path = '/sub-category';
@@ -17,15 +18,15 @@ class SubCategoryService {
     }
 
     save(form: SubCategorySaveDTO) {
-        return AxiosService.post(`${this.path}/save`, form)
+        return AxiosService.post(`${this.path}/save`, form, TokenService.getAuthorizationHeader())
     }
 
     update(form: SubCategory) {
-        return AxiosService.put(`${this.path}/save/${form.id}`, form)
+        return AxiosService.put(`${this.path}/save/${form.id}`, form, TokenService.getAuthorizationHeader())
     }
 
     delete(id: number) {
-        return AxiosService.delete(`${this.path}/${id}`)
+        return AxiosService.delete(`${this.path}/${id}`, TokenService.getAuthorizationHeader())
     }
     
 }
